cooperation_back/requirements.txt
python-dateutil
aiohttp
motor
sanic
pymongo
aioredis
sanic-auth
requests

